Separate samples on themes
