+++ samples/new_sample.py
@@ -1,0 +1,20 @@
+import json
+import os
+
+from mailjet_rest import Client
+
+mailjet30 = Client(auth=(os.environ["MJ_APIKEY_PUBLIC"],
+                         os.environ["MJ_APIKEY_PRIVATE"]))
+
+mailjet31 = Client(auth=(os.environ["MJ_APIKEY_PUBLIC"],
+                         os.environ["MJ_APIKEY_PRIVATE"]),
+                   version="v3.1")
+
+
+if __name__ == "__main__":
+    result = edit_contact_data()
+    print(result.status_code)
+    try:
+        print(json.dumps(result.json(), indent=4))
+    except json.decoder.JSONDecodeError:
+        print(result.text)
